+++ com.unity.visualeffectgraph/Editor/GraphView/Elements/VFXContextUI.cs
@@ -588,7 +588,7 @@ namespace UnityEditor.VFX.UI
 
             using (var growContext = new GrowContext(this))
             {
-                controller.AddBlock(blockIndex, descriptor.CreateInstance());
+                controller.AddBlock(blockIndex, descriptor.CreateInstance(), true /* freshly created block, should init space */);
             }
         }