+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Motion-Vector-Based Flipbook Interpolation
 - Color mapping mode for unlit outputs (Textured/Gradient Mapped)
 - Add Triangle and Octagon primitives for particle outputs
+- Set Attribute is now spaceable on a specific set of attributes (position, velocity, axis...)
 
 ### Fixed
 - Improve AA line rendering